+++ logprep/processor/list_comparison/processor.py
@@ -26,8 +26,6 @@ Processor Configuration
 .. automodule:: logprep.processor.list_comparison.rule
 """
 
-from ipaddress import ip_address
-
 from attr import define, field, validators
 
 from logprep.abc.processor import Processor
@@ -87,7 +85,7 @@ class ListComparison(Processor):
         field_value = get_dotted_field_value(event, rule.source_fields[0])
         value_list = field_value if isinstance(field_value, list) else [field_value]
 
-        # iterate over string lists and check if element is in any
+        # iterate over lists and check if element is in any
         list_matches = []
         for field_value in value_list:
             for compare_list in rule.compare_sets:
@@ -96,26 +94,7 @@ class ListComparison(Processor):
                 if field_value in rule.compare_sets[compare_list]:
                     list_matches.append(compare_list)
 
-        if rule.network_compare_sets:
-            for field_value in value_list:
-                self._compare_with_networks(list_matches, field_value, rule)
-
         # if matching list was found return it, otherwise return all list names
         if len(list_matches) == 0:
-            return rule.compare_set_keys, "not_in_list"
+            return list(rule.compare_sets.keys()), "not_in_list"
         return list_matches, "in_list"
-
-    @staticmethod
-    def _compare_with_networks(list_matches: list, field_value: str, rule: ListComparisonRule):
-        """Iterate over network lists, check if element is in any."""
-        try:
-            ip_address_object = ip_address(field_value)
-        except ValueError:
-            return
-
-        for compare_list in rule.network_compare_sets:
-            for network in rule.network_compare_sets[compare_list]:
-                if compare_list in list_matches:
-                    continue
-                if ip_address_object in network:
-                    list_matches.append(compare_list)