+++ tests/unit/ng/processor/list_comparison/test_list_comparison.py
@@ -1,6 +1,6 @@
 # pylint: disable=missing-docstring
 # pylint: disable=protected-access
-
+import pytest
 import responses
 
 from logprep.factory import Factory
@@ -288,3 +288,68 @@ Hans
         processor._rule_tree.add_rule(rule)
         processor.setup()
         assert processor.rules[0].compare_sets == {"bad_users.list": {"Franz", "Heinz", "Hans"}}
+
+    def test_list_comparison_does_not_add_duplicates_from_list_source(self):
+        document = {"users": ["Franz", "Alpha"]}
+        log_event = LogEvent(document, original=b"")
+        expected = {
+            "users": ["Franz", "Alpha"],
+            "user_results": {
+                "in_list": [
+                    "system_list.txt",
+                    "user_list.txt",
+                ]
+            },
+        }
+        rule_dict = {
+            "filter": "users",
+            "list_comparison": {
+                "source_fields": ["users"],
+                "target_field": "user_results",
+                "list_file_paths": [
+                    "../lists/system_list.txt",
+                    "../lists/user_list.txt",
+                ],
+            },
+            "description": "",
+        }
+        self._load_rule(rule_dict)
+        self.object.setup()
+        self.object.process(log_event)
+        assert document == expected
+
+    @pytest.mark.parametrize(
+        "testcase, system, result",
+        [
+            ("string in list", "Alpha", {"in_list": ["system_list.txt"]}),
+            ("string not in list", "Omega", {"not_in_list": ["system_list.txt"]}),
+            ("list element in list", ["Alpha"], {"in_list": ["system_list.txt"]}),
+            ("list element not in list", ["Omega"], {"not_in_list": ["system_list.txt"]}),
+            ("one list element in list", ["Alpha", "Omega"], {"in_list": ["system_list.txt"]}),
+            ("multiple list elements in list", ["Alpha", "Beta"], {"in_list": ["system_list.txt"]}),
+        ],
+    )
+    def test_match_list_field(self, testcase, system, result):
+        document = {"system": system}
+        log_event = LogEvent(document, original=b"")
+        expected = {"system": system, "system_results": result}
+        rule_dict = {
+            "filter": "system",
+            "list_comparison": {
+                "source_fields": ["system"],
+                "target_field": "system_results",
+                "list_file_paths": ["../lists/system_list.txt"],
+            },
+            "description": "",
+        }
+        config = {
+            "type": "ng_list_comparison",
+            "rules": [],
+            "list_search_base_path": self.CONFIG["list_search_base_path"],
+        }
+        processor = Factory.create({"custom_lister": config})
+        rule = processor.rule_class.create_from_dict(rule_dict)
+        processor._rule_tree.add_rule(rule)
+        processor.setup()
+        processor.process(log_event)
+        assert document == expected, testcase